+++ apps/resume/src/pages/resume/components/content-item/activity/activity-item.tsx
