+++ apps/resume/src/pages/resume/components/content-item/education/education-item.tsx
