+++ apps/resume/src/pages/resume/components/content-item/education/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import type { Education } from "../../../../../service";
 import { MotionStack } from "../../../../../components/stack";
-import EducationItem from "./EducationItem";
+import EducationItem from "./education-item";
 
 interface EducationSectionProps {
   education: Education[];
