+++ apps/resume/src/pages/resume/components/content-item/side-project/side-project-item.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import type { SideProject } from "../../../../../service";
 
-import { TimelineItem, type TimelineItemData } from "../shared/TimelineItem";
+import { TimelineItem, type TimelineItemData } from "../shared/timeline-item";
 
 /**
  * 사이드 프로젝트 아이템 컴포넌트
